Add serialization/deserialization for values. Fix #38
The Client is now able to serialize and deserialize values with
msgpack for every commands. This behavior can be turned off with the
Client.serialize flag or by passing 'serialize=False' to any
command.

The behavior of msgpack (use_list and use_bin_type) can also be
tweaked globally in the Client or locally in each command.